removed indirection from asset graph
