+++ crates/atlaspack_core/src/asset_graph/mod.rs
@@ -1,0 +1,8 @@
+#[allow(clippy::module_inception)]
+mod asset_graph;
+mod propagate_requested_symbols;
+mod serialize_asset_graph;
+
+pub use self::asset_graph::*;
+pub use self::propagate_requested_symbols::*;
+pub use self::serialize_asset_graph::*;
